// src/interfaces/leaveRequestFlow.interface.ts
export interface LeaveRequestFlow {
  currentStep?: string;
  companyCode: string;
  employeeCode?: string;
  leaveRequestDate?: Date;
  travelDate?: Date;
  leaveType?: string;
  leaveStartDate?: Date;
  leaveEndDate?: Date;
  leaveDays?: number;
  leaveReason?: string;
  daysAdjusted?: number;
  halfDay?: string;
  airTicket?: string;
  airTicketSelf?: string;
  airTicketWife?: string;
  airTicketChildren?: number;
  requestNumber: string;
  requestDate?: Date;
  flowCode?: string;
  flowLevelInitial?: number;
  flowLevelRunning?: number;
  flowLevelFinal?: number;
  faUploaded?: string;
  finalApproved?: string;
  createUser?: string;
  createDate?: Date;
  lastUpdated?: string;
  lastAction?: string;
  historySerial?: number;
  cancelFlag?: string;
  cancelUser?: string;
  cancelDate?: Date;
  cancelRemark?: string;
  remarksHistry?: string;
  remarks?: string;
  description?: string;
  comments?: string;
  mobileAppUpdate?: string;
  updatedAt?: Date;
  updatedBy?: string;
  createdBy?: string;
  createdAt?: Date;
  hod?: string;
  deptHead?: string;
  immediateSupervisor?: string;
  logNumber?: number;
  nextActionBy?: string;
  leaveAllowance?: string;
  advPayment?: string;
  causeType?: string;
  nameOfReplacement?: string;
  contactDetailsDuringLeave?: string;
  dutyResumeDate?: Date;
  actualResumeDate?: Date;
  employeeName?: string;
  dateFlag?: string;
}